+++ src/backend/order-service/src/main/java/com/spike/ticket/dto/request/CreateOrderRequest.java
@@ -3,7 +3,6 @@ package com.spike.ticket.dto.request;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -22,5 +21,5 @@ public class CreateOrderRequest {
     private List<Long> ticketIds;
 
     @NotNull
-    private BigDecimal totalPrice;
+    private Long totalPrice;
 }
