+++ src/backend/order-service/src/main/java/com/spike/ticket/dto/respone/OrderResponse.java
@@ -3,7 +3,6 @@ package com.spike.ticket.dto.respone;
 import lombok.Builder;
 import lombok.Data;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -12,6 +11,6 @@ public class OrderResponse {
     // id bên ngoài chỉ dùng tracking number, id đang là auto incre dễ lộ
     private String orderTrackingNumber;
     private String status;
-    private BigDecimal totalPrice;
+    private Long totalPrice;
     private LocalDateTime createdAt;
 }
