+++ src/backend/order-service/src/main/java/com/spike/ticket/dto/respone/TicketReservationResponse.java
@@ -5,13 +5,11 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
 public class TicketReservationResponse {
     private Long ticketId;
-    private BigDecimal price;    // Giá vé
+    private Long price;    // Giá vé
 }
