+++ src/backend/order-service/src/main/java/com/spike/ticket/entity/Order.java
@@ -7,7 +7,6 @@ import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
@@ -35,7 +34,7 @@ public class Order {
     private Long userId;
 
     @Column(name = "total_price", nullable = false, precision = 15, scale = 2)
-    private BigDecimal totalAmount;
+    private Long totalAmount;
 
     @Enumerated(EnumType.STRING)
     @Column(name = "status", nullable = false, length = 20)
